# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for FOC project not Frog OOB OOB.",
        "title": "FOC",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/ttf/create/draft": {
            "post": {
                "description": "crete draft document ttf.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "create draft",
                "operationId": "ttf-create-draft",
                "parameters": [
                    {
                        "description": "request body for create draft in JSON format",
                        "name": "create-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputCreateDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return from create draft",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/create/req": {
            "post": {
                "description": "create rquest document ttf.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "create request ttf",
                "operationId": "ttf-create-request",
                "parameters": [
                    {
                        "description": "request body for create draft in JSON format",
                        "name": "create-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputCreateDoc"
                        }
                    },
                    {
                        "description": "request body for create draft in JSON format",
                        "name": "create-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputCreateDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return from create draft",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/create/ret": {
            "post": {
                "description": "create return document ttf.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "create return ttf",
                "operationId": "ttf-create-return",
                "parameters": [
                    {
                        "description": "request body for create draft in JSON format",
                        "name": "create-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputCreateDoc"
                        }
                    },
                    {
                        "description": "request body for create draft in JSON format",
                        "name": "create-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputCreateDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return from create draft",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/edit/draft": {
            "post": {
                "description": "edit draft document ttf.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "edit draft",
                "operationId": "ttf-edit-draft",
                "parameters": [
                    {
                        "description": "request body for edit draft in JSON format",
                        "name": "edit-draft",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InputEditDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return from create draft",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/get-all": {
            "get": {
                "description": "Get all.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "Get all",
                "operationId": "ttf-get-all",
                "responses": {
                    "200": {
                        "description": "Get all document have been created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GetAllTTF"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/get-all/by": {
            "get": {
                "description": "Get all by operation's status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "Get all by operation's status",
                "operationId": "ttf-get-all-by-opr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation's status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all document by operation's status have been created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GetAllTTF"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/ttf/get-by": {
            "get": {
                "description": "Get by document's ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTF"
                ],
                "summary": "Get by document's ID",
                "operationId": "ttf-get-by",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document Number",
                        "name": "number",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all document by document's ID status have been created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GetTTFBy"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Api not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.EditTTFCustomer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "subDistrict": {
                    "type": "string"
                },
                "ttfNumber": {
                    "type": "string"
                }
            }
        },
        "dto.EditTTFDocument": {
            "type": "object",
            "properties": {
                "descs": {
                    "type": "string"
                },
                "descsRemark": {
                    "type": "string"
                },
                "locationFrom": {
                    "type": "string"
                },
                "locationTo": {
                    "type": "string"
                },
                "logisticType": {
                    "type": "integer"
                },
                "refOrder": {
                    "type": "string"
                },
                "reqType": {
                    "type": "string"
                },
                "siteFrom": {
                    "type": "string"
                },
                "siteTo": {
                    "type": "string"
                },
                "trackingNumber": {
                    "type": "string"
                },
                "ttfID": {
                    "type": "string"
                },
                "ttfNumber": {
                    "type": "string"
                },
                "userCreated": {
                    "type": "string"
                },
                "warehouseFrom": {
                    "type": "string"
                },
                "warehouseTo": {
                    "type": "string"
                }
            }
        },
        "dto.EditTTFItem": {
            "type": "object",
            "properties": {
                "descsItem": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "qtyOrder": {
                    "type": "integer"
                },
                "sizeID": {
                    "type": "integer"
                },
                "sizeName": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "skuName": {
                    "type": "string"
                },
                "ttfID": {
                    "type": "string"
                },
                "ttfNumber": {
                    "type": "string"
                }
            }
        },
        "dto.GetAllTTF": {
            "type": "object",
            "properties": {
                "head": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetHead"
                    }
                }
            }
        },
        "dto.GetCustomer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "subDistrict": {
                    "type": "string"
                },
                "ttfNumber": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetLine": {
            "type": "object",
            "properties": {
                "descsItem": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "qtyOrder": {
                    "type": "integer"
                },
                "sizeID": {
                    "type": "integer"
                },
                "sizeName": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "skuName": {
                    "type": "string"
                },
                "ttfID": {
                    "type": "string"
                },
                "ttfNumber": {
                    "type": "string"
                }
            }
        },
        "dto.GetTTFBy": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/dto.GetCustomer"
                },
                "head": {
                    "$ref": "#/definitions/dto.GetHead"
                },
                "line": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetLine"
                    }
                }
            }
        },
        "dto.InputCreateDoc": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/dto.TTFCustomer"
                },
                "head": {
                    "$ref": "#/definitions/dto.TTFDocument"
                },
                "line": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TTFItem"
                    }
                }
            }
        },
        "dto.InputEditDoc": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/dto.EditTTFCustomer"
                },
                "head": {
                    "$ref": "#/definitions/dto.EditTTFDocument"
                },
                "line": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EditTTFItem"
                    }
                }
            }
        },
        "dto.TTFCustomer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "subDistrict": {
                    "type": "string"
                }
            }
        },
        "dto.TTFDocument": {
            "type": "object",
            "properties": {
                "descs": {
                    "type": "string"
                },
                "descsRemark": {
                    "type": "string"
                },
                "locationFrom": {
                    "type": "string"
                },
                "locationTo": {
                    "type": "string"
                },
                "logisticType": {
                    "type": "integer"
                },
                "refOrder": {
                    "type": "string"
                },
                "reqType": {
                    "type": "string"
                },
                "siteFrom": {
                    "type": "string"
                },
                "siteTo": {
                    "type": "string"
                },
                "trackingNumber": {
                    "type": "string"
                },
                "userCreated": {
                    "type": "string"
                },
                "warehouseFrom": {
                    "type": "string"
                },
                "warehouseTo": {
                    "type": "string"
                }
            }
        },
        "dto.TTFItem": {
            "type": "object",
            "properties": {
                "descsItem": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "qtyOrder": {
                    "type": "integer"
                },
                "sizeID": {
                    "type": "integer"
                },
                "sizeName": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "skuName": {
                    "type": "string"
                }
            }
        },
...
    }
}